feat(Youtube api):  created a seprate function to fetch playlist data.
